Rename pkg/guest to pkg/vsphere/sys
