+++ lib/apiservers/engine/backends/cache/image_cache.go
@@ -34,7 +34,7 @@ import (
 	"github.com/vmware/vic/lib/apiservers/portlayer/client/storage"
 	"github.com/vmware/vic/lib/apiservers/portlayer/models"
 	"github.com/vmware/vic/lib/metadata"
-	"github.com/vmware/vic/pkg/guest"
+	"github.com/vmware/vic/pkg/vsphere/sys"
 )
 
 // ICache is an in-memory cache of image metadata. It is refreshed at startup
@@ -69,7 +69,7 @@ func (ic *ICache) Update(client *client.PortLayer) error {
 
 	log.Debugf("Updating image cache...")
 
-	host, err := guest.UUID()
+	host, err := sys.UUID()
 	if host == "" {
 		host, err = os.Hostname()
 	}
