+++ pkg/vsphere/sys/uuid_linux.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package guest
+package sys
 
 import (
 	"fmt"
