+++ pkg/vsphere/sys/uuid_other.go
@@ -14,7 +14,7 @@
 
 // +build !linux
 
-package guest
+package sys
 
 import "fmt"
