+++ pkg/vsphere/test/test.go
@@ -24,8 +24,8 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/vmware/govmomi/object"
 	"github.com/vmware/vic/lib/spec"
-	"github.com/vmware/vic/pkg/guest"
 	"github.com/vmware/vic/pkg/vsphere/session"
+	"github.com/vmware/vic/pkg/vsphere/sys"
 	"github.com/vmware/vic/pkg/vsphere/test/env"
 
 	"golang.org/x/net/context"
@@ -58,7 +58,7 @@ func Session(ctx context.Context, t *testing.T) *session.Session {
 
 // SpecConfig returns a spec.VirtualMachineConfigSpecConfig struct
 func SpecConfig(session *session.Session) *spec.VirtualMachineConfigSpecConfig {
-	uuid, err := guest.UUID()
+	uuid, err := sys.UUID()
 	if err != nil {
 		log.Errorf("unable to get UUID for guest - used for VM name: %s", err)
 		return nil
